rename location to site registry
