+++ netmanager/src/redux/SiteRegistry/actions.js
