+++ netmanager/src/redux/SiteRegistry/operations.js
