+++ netmanager/src/redux/SiteRegistry/reducers/index.js
