+++ netmanager/src/redux/SiteRegistry/reducers/location.js
