+++ netmanager/src/redux/SiteRegistry/selectors.js
