--- netmanager/src/views/components/LocationList/index.js
@@ -1,12 +1,0 @@
-export { default as LocationsTable }
-from "./LocationsTable";
-export { default as LocationsToolbar }
-from "./LocationsToolbar";
-export { default as LocationList }
-from "./LocationList";
-export { default as LocationEdit }
-from "./LocationEdit";
-export { default as LocationRegister }
-from "./LocationRegister";
-export { default as LocationView }
-from "./LocationView";
