+++ netmanager/src/views/components/SiteList/SiteList.js
@@ -4,7 +4,7 @@ import { makeStyles } from "@material-ui/styles";
 import { LocationsToolbar, LocationsTable } from "./index";
 import mockData from "./data";
 //import './assets/css/location-registry.css';
-import '../../../assets/css/location-registry.css';
+import "../../../assets/css/location-registry.css";
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -16,13 +16,13 @@ const useStyles = makeStyles((theme) => ({
   },
   title: {
     fontWeight: 700,
-    color: '#000000',
+    color: "#000000",
     fontSize: 24,
-    fontFamily: 'Open Sans'
+    fontFamily: "Open Sans",
   },
 }));
 
-const LocationList = () => {
+const SiteList = () => {
   const classes = useStyles();
 
   const [users] = useState(mockData);
@@ -32,11 +32,11 @@ const LocationList = () => {
       <LocationsToolbar />
       <div className={classes.content}>
         {/*<LocationsTable users={users} />*/}
-        <h5 className = {classes.title}> Location Registry </h5>
-        <LocationsTable/>
+        <h5 className={classes.title}> Site Registry </h5>
+        <LocationsTable />
       </div>
     </div>
   );
 };
 
-export default LocationList;
+export default SiteList;
