+++ netmanager/src/views/components/SiteList/SitesToolbar.js
@@ -5,7 +5,7 @@ import { makeStyles } from "@material-ui/styles";
 import { Button } from "@material-ui/core";
 import { Link } from "react-router-dom";
 import { SearchInput } from "../SearchInput";
-import {CSVLink, CSVDownload} from 'react-csv';
+import { CSVLink, CSVDownload } from "react-csv";
 
 const useStyles = makeStyles((theme) => ({
   root: {},
@@ -28,20 +28,20 @@ const useStyles = makeStyles((theme) => ({
     marginRight: theme.spacing(1),
   },
   link: {
-    color: '#3344FF',
+    color: "#3344FF",
     marginRight: theme.spacing(1),
-    fontWeight: "bold"
-    }
+    fontWeight: "bold",
+  },
 }));
 
-const LocationsToolbar = (props) => {
+const SitesToolbar = (props) => {
   const { className, ...rest } = props;
 
   const classes = useStyles();
 
-  let handleAddBtn = () =>{
-    window.location = '/register_location';
-  }
+  let handleAddBtn = () => {
+    window.location = "/register_location";
+  };
 
   return (
     <div {...rest} className={clsx(classes.root, className)}>
@@ -58,24 +58,24 @@ const LocationsToolbar = (props) => {
         >Export
         </Button>
       </CSVLink>*/}
-     <Link to="/register_location" >
-     <Button 
-          variant="contained" 
-          color="primary"              
-          type="submit"
-          align = "centre"
-        > Add Location
-        </Button>
-     </Link>    
-        
+        <Link to="/register_location">
+          <Button
+            variant="contained"
+            color="primary"
+            type="submit"
+            align="centre"
+          >
+            {" "}
+            Add Location
+          </Button>
+        </Link>
       </div>
-
     </div>
   );
 };
 
-LocationsToolbar.propTypes = {
+SitesToolbar.propTypes = {
   className: PropTypes.string,
 };
 
-export default LocationsToolbar;
+export default SitesToolbar;
