+++ netmanager/src/views/components/SiteList/index.js
@@ -1,0 +1,6 @@
+export { default as LocationsTable } from "./SitesTable";
+export { default as LocationsToolbar } from "./SitesToolbar";
+export { default as LocationList } from "./SiteList";
+export { default as LocationEdit } from "./SiteEdit";
+export { default as LocationRegister } from "./SiteRegister";
+export { default as LocationView } from "./SiteView";
